change float64 to float32, add AdaGrad, RMSProp, Adam

+++ autograd/tensor.py
@@ -12,7 +12,7 @@ def ensure_array(arrayable: Arrayable) -> np.ndarray:
     if isinstance(arrayable, np.ndarray):
         return arrayable
     else:
-        return np.array(arrayable,dtype=np.float64)
+        return np.array(arrayable,dtype=np.float32)
 
 Tensorable = Union['Tensor',float, np.ndarray]
 
@@ -48,7 +48,7 @@ class Tensor:
         self.grad = None
 
     def zero_grad(self) -> None:
-        self.grad = Tensor(np.zeros_like(self.data,dtype=np.float64))
+        self.grad = Tensor(np.zeros_like(self.data,dtype=np.float32))
 
     def __repr__(self) -> str:
         return f"Tensor({self.data}.requires_grad={self.requires_grad})"
